Clase 24/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
gradio
langchain-google-genai
python-dotenv

